Removing CHANGEME check conditions from stdout
After my first interview with @menecio, I stayed for a while playing with the quickstart repo.

### What?

I've seen that when you run the Makefile an **if condition** is being printed out (see the screenshot). I am not sure if it was intended, so I am opening this PR just in case.

I've also taken the liberty to:
- replace the `grep` with `findstring`, since it's a standard function within `GNU make`.
- standardize how strings are defined

Let me know what you think.